+++ pages/workspaces/[workspaceId]/repositories.tsx
@@ -1,8 +1,6 @@
 import { createPagesServerClient } from "@supabase/auth-helpers-nextjs";
 import { GetServerSidePropsContext } from "next";
-import { SquareFillIcon } from "@primer/octicons-react";
 import { useRouter } from "next/router";
-import { FaEdit } from "react-icons/fa";
 import { WorkspaceLayout } from "components/Workspaces/WorkspaceLayout";
 import { fetchApiData } from "helpers/fetchApiData";
 import Repositories from "components/organisms/Repositories/repositories";
@@ -17,6 +15,7 @@ import ClientOnly from "components/atoms/ClientOnly/client-only";
 import { deleteCookie } from "lib/utils/server/cookies";
 import { WORKSPACE_ID_COOKIE_NAME } from "lib/utils/workspace-utils";
 import Button from "components/atoms/Button/button";
+import { WorkspaceHeader } from "components/Workspaces/WorkspaceHeader";
 
 export const getServerSideProps = async (context: GetServerSidePropsContext) => {
   const supabase = createPagesServerClient(context);
@@ -58,20 +57,7 @@ const WorkspaceDashboard = ({ workspace }: WorkspaceDashboardProps) => {
 
   return (
     <WorkspaceLayout workspaceId={workspace.id}>
-      <section className="w-full flex justify-between items-center">
-        <h1 className="flex gap-2 items-center uppercase text-3xl font-semibold">
-          {/* putting a square icon here as a placeholder until we implement workspace logos */}
-          <SquareFillIcon className="w-12 h-12 text-sauced-orange" />
-          <span>{workspace.name}</span>
-        </h1>
-
-        <div className="flex gap-4 w-fit">
-          <Button variant="primary" href={`/workspaces/${workspace.id}/settings`} className="gap-2 items-center">
-            <FaEdit />
-            Edit
-          </Button>
-        </div>
-      </section>
+      <WorkspaceHeader workspace={workspace} />
       <div className="flex justify-between items-center">
         <WorkspacesTabList workspaceId={workspace.id} selectedTab={"repositories"} />
         <div className="flex items-center gap-4">