+++ pages/workspaces/[workspaceId]/settings.tsx
@@ -1,11 +1,11 @@
-import { FaEdit } from "react-icons/fa";
+import { FaRegCheckCircle } from "react-icons/fa";
 import { useRouter } from "next/router";
 import { GetServerSidePropsContext } from "next";
 import { createPagesServerClient } from "@supabase/auth-helpers-nextjs";
 import { ComponentProps, useState } from "react";
 import dynamic from "next/dynamic";
-import { SquareFillIcon } from "@primer/octicons-react";
 import { useEffectOnce } from "react-use";
+import { IoDiamond } from "react-icons/io5";
 import { WorkspaceLayout } from "components/Workspaces/WorkspaceLayout";
 import Button from "components/atoms/Button/button";
 import TextInput from "components/atoms/TextInput/text-input";
@@ -22,6 +22,7 @@ import {
   deleteTrackedRepos,
   deleteWorkspace,
   saveWorkspace,
+  upgradeWorkspace,
 } from "lib/utils/workspace-utils";
 import { WORKSPACE_UPDATED_EVENT } from "components/shared/AppSidebar/AppSidebar";
 import { WorkspacesTabList } from "components/Workspaces/WorkspacesTabList";
@@ -29,6 +30,10 @@ import { useGetWorkspaceContributors } from "lib/hooks/api/useGetWorkspaceContri
 import { TrackedContributorsTable } from "components/Workspaces/TrackedContributorsTable";
 import { deleteCookie } from "lib/utils/server/cookies";
 import WorkspaceVisibilityModal from "components/Workspaces/WorkspaceVisibilityModal";
+import Card from "components/atoms/Card/card";
+import { WorkspaceHeader } from "components/Workspaces/WorkspaceHeader";
+import { getStripe } from "lib/utils/stripe-client";
+
 
 const DeleteWorkspaceModal = dynamic(() => import("components/Workspaces/DeleteWorkspaceModal"), { ssr: false });
 
@@ -69,8 +74,15 @@ export const getServerSideProps = async (context: GetServerSidePropsContext) => 
     });
   }
 
+  if (!data?.members.find((member) => member.user_id === Number(sessionData?.id) && member.role === "owner")) {
+    return { notFound: true };
+  }
+
   return {
-    props: { workspace: data, canDeleteWorkspace: sessionData && workspaceId !== sessionData.personal_workspace_id },
+    props: {
+      workspace: data,
+      canDeleteWorkspace: sessionData && workspaceId !== sessionData.personal_workspace_id,
+    },
   };
 };
 
@@ -200,22 +212,21 @@ const WorkspaceSettings = ({ workspace, canDeleteWorkspace }: WorkspaceSettingsP
     }
   };
 
+  const upgradeThisWorkspace = async () => {
+    const { data, error } = await upgradeWorkspace({ workspaceId: workspace.id, sessionToken: sessionToken! });
+    if (error) {
+      toast({ description: "There's been an error", variant: "danger" });
+    }
+
+    if (data) {
+      const stripe = await getStripe();
+      stripe?.redirectToCheckout({ sessionId: data.sessionId as string });
+    }
+  };
+
   return (
     <WorkspaceLayout workspaceId={workspace.id}>
-      <section className="w-full flex justify-between items-center">
-        <h1 className="flex gap-2 items-center uppercase text-3xl font-semibold">
-          {/* putting a square icon here as a placeholder until we implement workspace logos */}
-          <SquareFillIcon className="w-12 h-12 text-sauced-orange" />
-          <span>{workspace.name}</span>
-        </h1>
-
-        <div className="flex gap-4 w-fit">
-          <Button variant="primary" href={`/workspaces/${workspace.id}/settings`} className="gap-2 items-center">
-            <FaEdit />
-            Edit
-          </Button>
-        </div>
-      </section>
+      <WorkspaceHeader workspace={workspace} />
       <div className="grid gap-6">
         <div>
           <div className="flex justify-between items-center">
@@ -296,15 +307,42 @@ const WorkspaceSettings = ({ workspace, canDeleteWorkspace }: WorkspaceSettingsP
             setTrackedContributorsPendingDeletion((contributors) => new Set([...contributors, contributor]));
           }}
         />
-
-        <div className="flex flex-col gap-4">
-          <Title className="!text-lg !leading-none py-6" level={4}>
-            Danger Zone
-          </Title>
+        
+        {workspace.payee_user_id ? (
+          <section className="flex flex-col gap-4">
+            <div className="flex gap-4 items-center">
+              <h3 className="font-medium">Manage Subscription</h3>
+              <div className="flex gap-2 items-center text-white px-3 py-2 bg-gradient-to-l from-gradient-orange-one to-gradient-orange-two rounded-full">
+                <p className="text-sm font-medium">PRO</p>
+                <IoDiamond />
+              </div>
+            </div>
+            <p className="text-sm text-slate-600">This Workspace is currently subscribed to the PRO Workspace plan.</p>
+            <Button href={process.env.NEXT_PUBLIC_STRIPE_SUB_CANCEL_URL} variant="primary" className="w-fit">
+              Manage Subscription
+            </Button>
+          </section>
+        ) : (
+          <Card className="flex flex-col gap-4 px-6 pt-5 pb-6">
+            <h2 className="text-md font-medium">Upgrade your workspace</h2>
+            <div className="flex gap-4">
+              <FaRegCheckCircle className="text-light-grass-8 w-6 h-6" />
+              <div className="flex flex-col gap-2">
+                <h3 className="text-sm font-medium">Make your workspace private</h3>
+                <p className="text-sm text-slate-500">
+                  Free workspaces can only be public, but with a pro workspace you can choose whether your workspace to
+                  be puclic or private!
+                </p>
+              </div>
+            </div>
+            <Button variant="primary" className="w-fit mt-2" onClick={upgradeThisWorkspace}>
+              Upgrade Workspace
+            </Button>
+          </Card>
+        )}
           <div className="flex flex-col p-6 rounded-2xl bg-light-slate-4">
             <Title className="!text-1xl !leading-none !border-light-slate-8 border-b pb-4" level={4}>
               Change Workspace Visibility
-            </Title>
             <Text className="my-4">This workspace is set to {isPublic ? "public" : "private"}.</Text>
 
             <Button onClick={() => setIsWorkspaceVisibilityModalOpen(true)} variant="destructive" className="w-fit">
@@ -325,8 +363,8 @@ const WorkspaceSettings = ({ workspace, canDeleteWorkspace }: WorkspaceSettingsP
                 </Button>
               </div>
             </div>
-          ) : null}
-        </div>
+          </div>
+        ) : null}
         <TrackedReposModal
           isOpen={trackedReposModalOpen}
           onClose={() => {